ML model/LLM/requirements.txt
# LLM Chatbot Requirements

fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.1
pymongo==4.10.1
google-generativeai==0.8.3
pydantic==2.9.2
websockets==13.1
httpx==0.27.2
